requirements.txt
Django==1.8
djangorestframework==3.4.0
pkg-resources==0.0.0
